Created package.json for backend and moved backend tests folder into backent. Also fixed the imports for every test.
